Removed the web.xml and mvc-dispatcher-servlet.xml since duplidates config
There seems to be somvc-dispatcher-servlet.xmlme confusion in that the web.xml configuration you
provided duplicates the WebAppInitializer you provided. Specifically
WebAppInitializer will configure the dispatcher servlet for you aready so
there is no need to configure it in the web.xml You can either configure the
DispatcherServlet in the web.xml OR configure it using WebAppInitializer
do not do both.

Additionally the mvc-dispatcher-servlet.xml was duplicating the configuration
of WebConfig. There is no need to do both. It is important to realize this
was creating two different DispatcherServlet instances. Each DispatcherServlet
isolates the Spring Beans so the one in the web.xml could see all beans
from the component scan. The other instance created by WebConfig was only
able to view beans created by WebConfig.java and those in teh root context.